requirements.txt
accelerate
albumentations
clip
clip-retrieval
datasets
deepspeed
diffusers==0.12.1
einops
ftfy
hydra-core
imageio==2.9.0
imageio-ffmpeg==0.4.2
Jinja2
kornia==0.6.11
lpips
openai-clip
opencv-python
pudb==2019.2
pytorch-lightning==1.4.2
tensorboard
test-tube>=0.7.5
torch-fidelity==0.3.0
torchmetrics==0.6.0
transformers==4.26.0
wandb
